批处理子模块：   1. 增加通过脚本参数输入功能   2. 增加delayMs，以延时看清编译过程情况   3. 上述修改更新批处理脚本，并调试测试完成！

+++ Pro_CfgCompile.cpp
@@ -132,7 +132,7 @@ int Dialog::Pro_CfgCompileData(QDataStream &dest, QString &Desc, QString &Data)
     }
     else if(Desc == "HEX64"){
       if(bh64 == false) Len = -1; //变量值表达错误
-      else{ dest << (quint64)h64; Len = 8; }
+      else{ dest << h64; Len = 8; }
     }
     else if(Desc == "RGB"){
       if(Data[0] != '#') Len = -1; //变量值表达错误
